fix(eventbridge): 8 AWS-spec divergences + cron L/W/# parity
Added:
- cron L / LW / <n>W / <n>L / <n>#<k> operators + DoM/DoW mutual-exclusion at PutRule

Fixed:
- DescribeRule / ListRules emit CreatedBy and ManagedBy (spec members dropped)
- PutEvents rejects >10 entries (AWS spec cap)
- event Time is int epoch (was float; Java/Go SDK v2 reject) — also affects archive replay
- content-filter [{exists: false}] now matches absent keys
- ListRules paginates (Limit + opaque NextToken)
- ListRuleNamesByTarget NextToken is opaque (was raw integer offset)
- DescribeEventBus / ListEventBuses omit Policy when unset (was empty string)
- DescribeEventSource State returns ACTIVE (was invalid 'ENABLED'; enum is PENDING/ACTIVE/DELETED)

AppSync Events:
- Lambda authorizer payload rewritten to AWS shape (authorizationToken + requestContext{apiId,accountId,requestId,operation,channelNamespaceName,channel} + requestHeaders); operations EVENT_CONNECT/EVENT_SUBSCRIBE/EVENT_PUBLISH; response field handlerContext (was resolverContext)
- ChannelNamespace response includes channelNamespaceArn (spec member)
- Channel grammar leading slash is optional (was required)
- Max 5 events per publish enforced (HTTP + WebSocket)
- Duplicate subscribe id returns subscribe_error (was silent overwrite)
- Unsubscribe on unknown id returns unsubscribe_error/UnknownOperationError
- ListApis / ListChannelNamespaces / ListApiKeys paginate (maxResults + nextToken)
- logConfig omitted when unset (was null)

CloudFront KVS:
- data-plane added to _state_map (PERSIST_STATE was dropping keys on restart)
- DescribeKeyValueStore Created / LastModified parsed from mgmt timestamp (was 0)
- UpdateKeys is atomic (validate all then commit; rejects whole batch on bad entry)
- ListKeys NextToken is opaque base64 (was raw key name)
- ListKeys MaxResults capped at 50 per AWS spec
- DeleteKey returns ResourceNotFoundException on missing key (was silent 200)
- s3 vhost bypass via _NON_S3_VHOST_NAMES (cloudfront-kvs) instead of path-prefix check

App / logging:
- per-service 'init completed' lines demoted to DEBUG (startup avalanche)
- INFO logs on resource creation: S3 bucket, DynamoDB table, SQS queue, SNS topic, Lambda function

Tests:
- AppSync test files consolidated into tests/test_appsync.py (was 3 files)
- cron L/W/#/mutual-exclusion + EventBridge fixes covered

CHANGELOG:
- 1.3.23 cron line restored to its original (rate-only) form; full cron + the 8 fixes moved to [Unreleased]
